Fixed delegation UI issues
This patch fixes several issues in delegation UI:

When adding a new delegation, only the first attribute selected
was saved. Now all attributes will be saved properly.

When loading the details page, the custom widgets did not store
the original values properly so is_dirty() did not work correctly.
Now this has been fixed except for the memberof attribute because
of these issues:
 - https://fedorahosted.org/freeipa/ticket/869
 - https://fedorahosted.org/freeipa/ticket/870

When saving the details page, the attrs were saved as an array
which was rejected by the server. Now it is stored as comma-
separated list.